Moved deactivate/reactivate functionality to committee model.
